requirements.txt
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.6
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-import-export==2.5.0
django-js-asset==1.2.2
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.1.0
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
